refactor(tui): swap hand-typed types.ts for generated types.generated.ts
Completes the TUI spec's "Option A -> Option B" promotion path.
The hand-typed types.ts (written in deliberately codegen-shaped form
during #489) is replaced by tui/src/types.generated.ts, produced by
scripts/gen_message_types.py from the same message_types.json
manifest that drives the Python side.

Consumer changes:
- 4 import paths updated: `./types.js` -> `./types.generated.js`
  (wsClient.ts, dispatcher.ts, dispatcher.test.ts, App.tsx)
- conv_selected case loses `model` assignment (the field moved to
  conv_history.active_model in the manifest reconciliation)
- conv_history case now wires `m.active_model ?? s.model` into
  state.model — the connection-time state dump is the earliest point
  the TUI can learn the active model
- 3 small test fixture updates to match the corrected wire shapes:
  compaction_done gains before/after counts, error.conv_id is now
  optional rather than nullable, conv_selected drops model assertion

The 12 vitest dispatcher tests pass unchanged in semantics. The
generated file's drift check is part of `make check-message-types`,
so the TUI's wire types can't silently rot.

Closes #492.
